pyproject.toml
[tool.poetry]
name = "mack"
version = "0.1.0"
description = ""
authors = ["Matthew Powers <matthewkevinpowers@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
pyspark = "3.3.1"
delta-spark = "2.1.1"
pytest = "3.2.2"
chispa = "0.9.2"
pytest-describe = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
